Issue #111 Refactor task detail connected component after rebase
